Nova Launcher - Get rid of "warning: modified apk" message in menu
Same version, added libs back into apk file (previously removed
and put in prebuilt libs). This wasn't an issue in version 2.xx
because there were no native C++ libs to the apk until version 3.
I've tried resigning it, and even removing the lines for the 4
libs from the META-INF/MANIFEST.MF file manually, but the warning
message persists. This is the only way I've been able to get rid
of it.

Change-Id: I98ccc31c5112c441998db1c5707c601a65f3391a